Move old file out
